# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Concept Back Exo API",
    "version": "1.0.0",
    "description": "Documentation générée par swagger-jsdoc"
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123"
          },
          "name": {
            "type": "string",
            "example": "Alice"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "alice@example.com"
          }
        },
        "required": [
          "id"
        ]
      },
      "UserCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Alice"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "alice@example.com"
          }
        },
        "required": [
          "name",
          "email"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 404
          },
          "message": {
            "type": "string",
            "example": "User not found"
          },
          "details": {
            "type": "string",
            "example": "Additional error details (optional)"
          }
        },
        "required": [
          "message"
        ]
      },
      "securitySchemes": {
        "BearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  },
  "paths": {
    "/hello": {
      "get": {
        "summary": "Retourne un message de salutation",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Nom optionnel pour personnaliser le message",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Bonjour, Alice !"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get user by ID",
        "description": "Récupère les informations d'un utilisateur donné son identifiant. Nécessite un token Bearer.",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant de l'utilisateur",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Objet utilisateur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "id": "123",
                  "name": "Alice",
                  "email": "alice@example.com"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid token",
            "content": {
              "application/json": {
                "example": {
                  "code": 401,
                  "message": "Unauthorized",
                  "details": "Missing or invalid Authorization header"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "example": {
                  "code": 404,
                  "message": "User not found",
                  "details": "No user with id 0"
                },
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "post": {
        "summary": "Create user",
        "description": "Crée un nouvel utilisateur à partir d'un nom et d'un email. Retourne l'utilisateur créé avec son identifiant.",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              },
              "example": {
                "name": "Bob",
                "email": "bob@example.com"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur créé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "id": "1762336349302",
                  "name": "Bob",
                  "email": "bob@example.com"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}